Refactor MSBuild environment setup to a separate class
